Refactor to TypeScript and modularize dashboard
Migrated App from JSX to TypeScript and moved weather and quote logic into separate components under src/components. Updated main.tsx to use the new App.tsx and added FontAwesome CSS import. Removed obsolete weather.tsx and App.jsx files, and renamed quote.tsx to components/quote.tsx for better organization.
