--- src/App.jsx
@@ -1,15 +1,0 @@
-import "./App.css";
-import "./weather"
-import "./dailyQuote"
-
-
-function App() {
-  return (
-    <div className="bg-amber-700 min-h-screen bg-primary flex flex-col items-center justify-center text-white font-sans">
-      <h1 className="text-4xl">Welcome to the Mini Dashboard 🚀</h1>
-      
-    </div>
-  );
-}
-
-export default App;
