+++ src/App.tsx
@@ -1,0 +1,15 @@
+
+import WeatherCard from "./components/weather.tsx";
+
+
+function App() {
+  return (
+    <div className="min-h-screen bg-gray-900 py-8 flex flex-col md:flex-row gap-4 justify-center items-start">
+      
+      <WeatherCard />
+  
+    </div>
+  );
+}
+
+export default App;
